requirements.txt
Flask==3.1.2
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.9
MarkupSafe==3.0.3
PyYAML==6.0.3
Pygments==2.19.2
SQLAlchemy==2.0.44
TA-Lib==0.6.8
Unidecode==1.4.0
Werkzeug==3.1.3
aiosqlite==0.21.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apprise==1.9.5
asgi-lifespan==2.1.0
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
coolname==2.2.0
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
debugpy==1.8.17
decorator==5.2.1
docker==7.1.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.120.2
fonttools==4.60.1
frozendict==2.4.6
fsspec==2025.9.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2-humanize-extension==0.4.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
matplotlib==3.10.7
mdurl==0.1.2
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.3.1
opentelemetry-api==1.38.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathlib==1.0.1
pathspec==0.12.1
peewee==3.18.2
pendulum==3.1.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prefect==3.4.25
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
pyzmq==27.1.0
readchar==4.2.1
referencing==0.37.0
regex==2025.10.23
requests-oauthlib==2.0.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml.clib==0.2.14
ruamel.yaml==0.18.16
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.49.1
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.9.6
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
xgboost-cpu==3.1.1
yfinance==0.2.66
zipp==3.23.0
